Bluetooth: Mesh: Generic OnOff & Level Models demo
Removed process_message.c & process_message.h
& merge its content into respective handler functon
in device_composition.c. Plus some minor improvements.
