+++ samples/boards/nrf52/mesh/onoff-app/prj.conf
@@ -2,14 +2,6 @@
 CONFIG_MAIN_STACK_SIZE=512
 CONFIG_SYSTEM_WORKQUEUE_STACK_SIZE=2048
 
-CONFIG_BT_SETTINGS=y
-CONFIG_FLASH=y
-CONFIG_FLASH_PAGE_LAYOUT=y
-CONFIG_FLASH_MAP=y
-CONFIG_FCB=y
-CONFIG_SETTINGS=y
-CONFIG_SETTINGS_FCB=y
-
 CONFIG_BOOT_BANNER=y
 CONFIG_BUILD_TIMESTAMP=y
 
@@ -68,7 +60,7 @@ CONFIG_SYS_LOG_DEFAULT_LEVEL=4
 CONFIG_SYS_LOG_EXT_HOOK=y
 CONFIG_BT_DEBUG_LOG=y
 
-#CONFIG_BT_MESH_DEBUG=y
+CONFIG_BT_MESH_DEBUG=y
 #CONFIG_BT_MESH_DEBUG_PROV=y
 #CONFIG_BT_MESH_DEBUG_PROXY=y
 #CONFIG_BT_MESH_DEBUG_BEACON=y
@@ -104,5 +96,3 @@ CONFIG_BT_ATT_PREPARE_COUNT=2
 #CONFIG_SYS_LOG_NET_BUF_LEVEL=4
 #CONFIG_NET_BUF_LOG=y
 #CONFIG_BT_CTLR_DEBUG_PINS=y
-
-#CONFIG_BT_MESH_RPL_STORE_TIMEOUT=5
