+++ samples/boards/nrf52/mesh/onoff_level_app/CMakeLists.txt
@@ -11,7 +11,6 @@ target_sources(app PRIVATE
 	       src/mesh/ble_mesh.c
                src/mesh/device_composition.c
                src/mesh/publisher.c
-               src/mesh/process_message.c
 	       )
 
 zephyr_include_directories(
