+++ samples/boards/nrf52/mesh/onoff_level_app/src/common.h
@@ -28,7 +28,7 @@
 #include "ble_mesh.h"
 #include "device_composition.h"
 #include "publisher.h"
-#include "process_message.h"
+//#include "process_message.h"
 
 /* -------------------- COMPANY ID -------------------- */
