--- samples/boards/nrf52/mesh/onoff_level_app/src/mesh/process_message.h
@@ -1,9 +1,0 @@
-#ifndef _DOER_H
-#define _DOER_H
-
-void process_message(struct bt_mesh_model *model,
-		     struct bt_mesh_msg_ctx *ctx,
-		     struct net_buf_simple *buf,
-		     uint16_t opcode);
-
-#endif
